Rename hammy.js -> handler.js, following serverless convention
